Struct oas_common::record::TypedRecord

pub struct TypedRecord<T> where
    T: Clone,  {
    pub meta: RecordMeta,
    pub value: T,
}

A record with a strongly typed value.

All values should implement TypedValue.

Fields
meta: RecordMetavalue: T
Implementations

impl<T> TypedRecord<T> where
T: TaskObject,

pub fn task_states(&self) -> Option<&<T as TaskObject>::TaskStates>

pub fn task_states_mut(&mut self) -> Option<&mut <T as TaskObject>::TaskStates>

impl<T> TypedRecord<T> where
T: TypedValue,

pub fn guid(&self) -> &str

Get the guid of the record.

pub fn id(&self) -> &str

Get the id of the record.

pub fn typ(&self) -> &str

Get the typ of the record.

pub fn from_id_and_value(id: impl ToString, value: T) -> Self

Create a new record from an id and a value.

pub fn into_untyped_record(self) -> Result<UntypedRecord, EncodingError> where
T: Serialize,

Convert this record into an UntypedRecord.

This can be unwrapped by default as it only fails if the record value would not serialize to an object (which should be treated as a bug).

pub fn into_json_object(self) -> Result<Object, EncodingError>

Convert this record into a JSON Object.

pub fn validate(&self) -> Result<(), ValidationError>

Validate the contained value (according to the TypedValue implementation)

impl<T> TypedRecord<T> where
T: Resolvable + Send,

pub async fn resolve_refs<R: Resolver + Send + Sync>(
&mut self,
resolver: &R
) -> Result<(), MissingRefsError>

Resolve all references within this record into loaded records.

pub async fn into_resolve_refs<R: Resolver + Send + Sync>(
self,
resolver: &R
) -> Result<Self, MissingRefsError>

Resolve all references, while consuming Self and returning it again with the resolved refs.

pub fn extract_refs(&mut self) -> Vec<UntypedRecord>

Extract all loaded referenced records from within the record, converting the references back to IDs.