Struct oas_common::record::UntypedRecord

pub struct UntypedRecord { /* private fields */ }

An untyped record is a record without static typing. The value is encoded as a JSON object.

Implementations

impl UntypedRecord

pub fn with_typ_id_value(
typ: &str,
id: &str,
value: Value
) -> Result<Self, DecodingError>

Create an untyped record from type and id strings and a serde_json::Value.

The value should be a serde_json::Map, otherwise this method will return an DecodingError::NotAnObject.

pub fn into_typed_record<T: TypedValue + DeserializeOwned + Clone + 'static>(
self
) -> Result<TypedRecord<T>, DecodingError>

pub fn into_typed<T: TypedValue + DeserializeOwned + Clone + 'static>(
self
) -> Result<TypedRecord<T>, DecodingError>

Convert this untyped record into a typed [Record]. #[deprecated(note = “use into_typed”)]

pub fn into_json_object(self) -> Result<JsonObject, EncodingError>

Convert the untyped record into a JSON [Object].

pub fn guid(&self) -> &str

Get the guid of the record.

pub fn id(&self) -> &str

Get the id of the record.

pub fn typ(&self) -> &str

Get the type of the record.

pub fn meta(&self) -> &RecordMeta

pub fn meta_mut(&mut self) -> &mut RecordMeta

pub fn merge_json_value(
&mut self,
value_to_merge: Value
) -> Result<(), EncodingError>

Merge this record’s value with another JSON value.

pub fn apply_json_patch(&mut self, patch: &Patch) -> Result<(), EncodingError>