Trait oas_common::record::TypedValue

pub trait TypedValue: Debug + Any + Serialize + DeserializeOwned + Clone + Send {
    const NAME: &'static str;
    fn label(&self) -> Option<&str> { ... }

    fn guid(id: &str) -> String { ... }

    fn validate(&self) -> Result<(), ValidationError> { ... }
}

A trait to implement on value structs for typed Records.

Associated Constants

const NAME: &'static str

A string to uniquely identify this record type.

Provided methods

fn label(&self) -> Option<&str>

Get a human-readable label for this record.

This method is optional and returns None by default. Record types may implement this method to return a title or headline.

fn guid(id: &str) -> String

Get the guid string for this record type and an id string.

fn validate(&self) -> Result<(), ValidationError>