Enum oas_common::reference::Reference

pub enum Reference<T: Clone> {
    Id(String),
    Resolved(Record<T>),
}

A reference is an enum that can be in two states: It either holds an ID to a record, or a loaded record.

Variants

Id(String)

Resolved(Record<T>)

Implementations

impl<T: TypedValue> Reference<T>

pub fn id(&self) -> &str

Get the target id of the reference.

pub fn guid(&self) -> &str

pub fn record(&self) -> Option<&Record<T>>

Get the loaded record if the reference is resolved.

pub fn record_mut(&mut self) -> Option<&mut Record<T>>

Get a mutable reference to the loaded record if the reference is resolved.

pub fn into_record(self) -> Option<Record<T>>

Get the loaded record if the reference is resolved, while consuming the reference.

pub fn extract_record(&mut self) -> Option<Record<T>>

Extract the loaded record if the reference is resolved, while converting the reference to only hold an id.

pub fn resolved(&self) -> bool

Check if the reference is resolved and contains a loaded record.

pub async fn resolve<R: Resolver>(
&mut self,
resolver: &R
) -> Result<(), ResolveError>

Resolve this reference with a Resolver if the reference is not yet in the resolved state.

The Resolver trait is implemented on data stores.

pub async fn force_resolve<R: Resolver>(
&mut self,
resolver: &R
) -> Result<(), ResolveError>

Resolve the record, even if the reference is already in the resolved state.

pub async fn into_resolved<R: Resolver>(
self,
resolver: &R
) -> Result<Self, ResolveError>

Resolve the record while consuming self, and return a new reference in the resolved state.

pub fn set_resolved(&mut self, record: Record<T>)

Set the record to the resolved state while passing in the target record.